+++ CHANGES.rst
@@ -21,6 +21,7 @@ New features:
 
 Bug fixes:
 
+- Fix ``STANDARD`` and ``DAYLIGHT`` components that have a date as ``DTSTART``. See `Issue 218 <https://github.com/collective/icalendar/issues/218>`_
 - Move import at the end of ``icalendar.parser`` into a function to mitigate import errors, see `Issue 781 <https://github.com/collective/icalendar/issues/781>`_.
 
 6.1.3 (2025-03-19)